+++ rethinkdb-net/Query.cs
@@ -450,6 +450,11 @@ namespace RethinkDb
             return new NthQuery<T>(sequenceQuery, index);
         }
 
+        public static IsEmptyQuery<T> IsEmpty<T>(this ISequenceQuery<T> sequenceQuery)
+        {
+            return new IsEmptyQuery<T>(sequenceQuery);
+        }
+
         public static DistinctQuery<T> Distinct<T>(this ISequenceQuery<T> sequenceQuery)
         {
             return new DistinctQuery<T>(sequenceQuery);